add comments

+++ compare_genome_assemblies/convert.BED.chrNames.py
@@ -17,6 +17,9 @@ Correspondence = path + "data/genome_sequences/" + species + "/chromosome_corres
 MatrixPath = path + "results/substitution_matrix/" + species
 chromosomes_list = [chrom.strip('.txt') for chrom in os.listdir(MatrixPath)]
 
+if not os.path.isfile(Correspondence):
+    sys.exit("Chromosome correspondence file doesn't exist!")
+
 output = open(BED + "_UCSC_names", 'w')
 
 ####################################################################################################
